+++ main/index.js
@@ -1,5 +1,5 @@
-const {writeFile} = require('fs/promises')
-const {dirname, join} = require('path')
+const {readFile, writeFile} = require('fs/promises')
+const {basename, dirname, join} = require('path')
 const {app, BrowserWindow, dialog, ipcMain, Menu, shell} = require('electron')
 const {default: Conf} = require('conf')
 const Papa = require('papaparse')
@@ -9,7 +9,7 @@ const {PropertyType} = require('../lib/uexport')
 
 const UPACKAGE_OPEN_DIALOG_DEFAULT_PATH_ID = 'upackageOpenDialogDefaultPath'
 const UPACKAGE_SAVE_DIALOG_DEFAULT_PATH_ID = 'upackageSaveDialogDefaultPath'
-const CSV_SAVE_DIALOG_DEFAULT_PATH_ID = 'csvSaveFileDialogDefaultPath'
+const CSV_DIALOG_DEFAULT_PATH_ID = 'csvFileDialogDefaultPath'
 
 const conf = new Conf()
 
@@ -63,6 +63,13 @@ function createMainWindow() {
           click: saveUPackage,
         },
         {type: 'separator'},
+        {
+          label: 'Import...',
+          id: 'import',
+          enabled: false,
+          accelerator: 'Control+I',
+          click: importCSV,
+        },
         {
           label: 'Export...',
           id: 'export',
@@ -146,6 +153,7 @@ async function openUPackage() {
       mainWindow.webContents.send('upackage-opened', JSON.stringify(upackage))
       const menu = Menu.getApplicationMenu()
       menu.getMenuItemById('save').enabled = true
+      menu.getMenuItemById('import').enabled = true
       menu.getMenuItemById('export').enabled = true
       conf.set(UPACKAGE_OPEN_DIALOG_DEFAULT_PATH_ID, dirname(filePaths[0]))
     } catch (err) {
@@ -214,6 +222,183 @@ async function exportCSV() {
   mainWindow.webContents.send('export-csv')
 }
 
+const csvArrayFieldRegExp = /^(.+)\[(\d+)]$/
+
+async function importCSV() {
+  try {
+    const {canceled, filePaths} = await dialog.showOpenDialog({
+      defaultPath: conf.get(CSV_DIALOG_DEFAULT_PATH_ID),
+      filters: [
+        {name: 'CSV files', extensions: ['csv']},
+        {name: 'All files', extensions: ['*']},
+      ],
+    })
+
+    if (!canceled) {
+      const csv = await readFile(filePaths[0], 'utf8')
+      const {data, errors} = Papa.parse(csv, {
+        delimiter: ',',
+        header: true,
+        skipEmptyLines: true,
+      })
+
+      if (errors.length > 0) {
+        throw errors[0]
+      }
+
+      const {uasset, uexp} = upackage
+      const {props, entries} = uexp
+      const csvBasename = basename(filePaths[0])
+      const uassetBasename = basename(uasset.filename)
+      const uexpBasename = basename(uexp.filename)
+      const csvEntries = []
+
+      if (data.length !== entries.length) {
+        throw new Error(
+          `The number of records in ${csvBasename} does not match the number of entries in ${uexpBasename}. Expected ${entries.length}, got ${data.length}`,
+        )
+      }
+
+      for (let i = 0; i < data.length; i++) {
+        const line = data[i]
+        const entry = entries[i]
+        const lineNumber = i + 2
+        const csvEntry = {}
+
+        for (const field of Object.keys(line)) {
+          const value = line[field]
+          if (field === 'Tag') {
+            if (line.Tag !== entry.$tag) {
+              throw new Error(
+                `Tag mismatch in ${csvBasename} on line ${lineNumber}. Expected '${entry.$tag}', got '${line.Tag}'`,
+              )
+            }
+
+            csvEntry.$tag = value
+          } else {
+            const arrayMatch = csvArrayFieldRegExp.exec(field)
+            const propName = arrayMatch != null ? arrayMatch[1] : field
+
+            const prop = props.find(prop => prop.name === propName)
+            if (prop == null) {
+              throw new Error(`Unknown field '${propName}' in ${csvBasename}`)
+            }
+
+            const entryValue = entry[propName]
+            if (arrayMatch != null) {
+              if (csvEntry[propName] == null) {
+                csvEntry[propName] = []
+              }
+
+              const index = Number(arrayMatch[2])
+              if (index >= entryValue.length) {
+                if (value === '') {
+                  continue
+                }
+
+                throw new RangeError(
+                  `Array index out of range for field ${field} in ${csvBasename} on line ${lineNumber} for corresponding entry in ${uexpBasename}. Expected max ${
+                    entryValue.length - 1
+                  }, got ${index}`,
+                )
+              }
+
+              const entryElement = entryValue[index]
+              let number
+              switch (prop.type) {
+                case PropertyType.BOOLEAN:
+                case PropertyType.BYTE:
+                case PropertyType.BOOLEAN_BYTE:
+                case PropertyType.UINT16:
+                case PropertyType.INT32:
+                case PropertyType.FLOAT:
+                  number = Number(value)
+                  if (isNaN(number)) {
+                    throw new TypeError(
+                      `Invalid value for field ${field} in ${csvBasename} on line ${lineNumber}. Expected a number, got '${value}'`,
+                    )
+                  }
+
+                  csvEntry[propName][index] = number
+                  break
+                case PropertyType.STRING:
+                  if (value !== entryElement) {
+                    throw new Error(
+                      `Unsupported string modification for field ${field} in ${csvBasename} on line ${lineNumber}`,
+                    )
+                  }
+
+                  csvEntry[propName][index] = value
+                  break
+                case PropertyType.NAME:
+                  if (!uasset.names.includes(value)) {
+                    throw new RangeError(
+                      `Invalid name '${value}' for field ${field} in ${csvBasename} on line ${lineNumber}. The name does not exist in ${uassetBasename}`,
+                    )
+                  }
+
+                  csvEntry[propName][index] = value
+                  break
+                default:
+                  throw new Error(
+                    `Unsupported property type ${prop.type} in ${uexpBasename}`,
+                  )
+              }
+            } else {
+              let number
+              switch (prop.type) {
+                case PropertyType.BOOLEAN:
+                case PropertyType.BYTE:
+                case PropertyType.BOOLEAN_BYTE:
+                case PropertyType.UINT16:
+                case PropertyType.INT32:
+                case PropertyType.FLOAT:
+                  number = Number(value)
+                  if (isNaN(number)) {
+                    throw new TypeError(
+                      `Invalid value for field ${field} in ${csvBasename} on line ${lineNumber}. Expected a number, got '${value}'`,
+                    )
+                  }
+
+                  csvEntry[propName] = number
+                  break
+                case PropertyType.STRING:
+                  if (value !== entryValue) {
+                    throw new Error(
+                      `Unsupported string modification for field ${field} in ${csvBasename} on line ${lineNumber}`,
+                    )
+                  }
+
+                  csvEntry[propName] = value
+                  break
+                case PropertyType.NAME:
+                  if (!uasset.names.includes(value)) {
+                    throw new RangeError(
+                      `Invalid name '${value}' for field ${field} in ${csvBasename} on line ${lineNumber}. The name does not exist in ${uassetBasename}`,
+                    )
+                  }
+
+                  csvEntry[propName] = value
+                  break
+                default:
+                  throw new Error(
+                    `Unsupported property type ${prop.type} in ${uexpBasename}`,
+                  )
+              }
+            }
+          }
+        }
+
+        csvEntries[i] = csvEntry
+      }
+
+      mainWindow.webContents.send('csv-imported', csvEntries)
+    }
+  } catch (err) {
+    dialog.showMessageBoxSync({message: err.stack})
+  }
+}
+
 ipcMain.on('csv-exported', (event, entries) => {
   csvExported(entries)
 })
@@ -277,7 +462,7 @@ async function csvExported(entries) {
     const csv = Papa.unparse({fields, data})
     const {canceled, filePath} = await dialog.showSaveDialog({
       defaultPath:
-        conf.get(CSV_SAVE_DIALOG_DEFAULT_PATH_ID) ||
+        conf.get(CSV_DIALOG_DEFAULT_PATH_ID) ||
         uexp.filename.replace(/\.uexp$/, '.csv'),
       filters: [
         {name: 'CSV files', extensions: ['csv']},
@@ -287,7 +472,7 @@ async function csvExported(entries) {
 
     if (!canceled) {
       await writeFile(filePath, csv)
-      conf.set(CSV_SAVE_DIALOG_DEFAULT_PATH_ID, filePath)
+      conf.set(CSV_DIALOG_DEFAULT_PATH_ID, filePath)
     }
   } catch (err) {
     dialog.showMessageBoxSync({message: err.stack})